fix(pico-page): extract pure logic so test doesn't load runtime config.json
picoDisplay.tsx imports ../config.json (runtime-generated, gitignored, not
present in CodeBuild at synth-stage test time). The test imported from
picoDisplay.tsx, which transitively triggered the config.json import and
failed: 'Cannot find module ../config.json'.

Extract types + generateConfig into picoDisplay.config.ts (no runtime
imports). Test imports from .config.ts; component re-exports the symbols
so external callers keep working. Same pattern App.tsx avoids by being
in the vitest exclude list — picoDisplay.test.tsx is included so it needs
to load cleanly.
